# cnf/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Classic Auctions",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "produces": ["application/json"],
...
    "paths": {
...
        "/regions": {
            "get": {
                "tags": ["regions"],
                "responses": {
                    "200": { "description": "OK" },
                    "400": { "description": "Bad Request" },
                    "401": { "description": "Unauthorized" }
                },
                "parameters": []
            },
            "post": {
                "tags": ["regions"],
                "responses": {
                    "200": { "description": "OK" },
                    "400": { "description": "Bad Request" },
                    "401": { "description": "Unauthorized" }
                },
                "parameters": [
                    {
                        "name": "short_name",
                        "in": "query",
                        "description": "Region name abbreviation",
                        "type": "string"
                    },
                    {
                        "name": "long_name",
                        "in": "query",
                        "description": "Region full name",
                        "type": "string"
                    }
                ]
            }
        },
...
    }
}